+++ src/modules/core/Color.h
@@ -82,7 +82,7 @@ public:
 	}
 	static glm::vec4 fromRGBA(uint8_t r, uint8_t g, uint8_t b, uint8_t a = 255);
 	static glm::vec4 fromHSB(const float hue, const float saturation, const float brightness, const float alpha = 1.0f);
-	static glm::vec4 fromHex(const char* hex);
+	static core::RGBA fromHex(const char* hex);
 	static core::String toHex(const RGBA rgba, bool hashPrefix = true);
 
 	static RGBA getRGBA(const glm::vec4&);